Revert incorrect AWS SK update for v1 modules Revert accidental package renaming
