+++ interlok-aws-kinesis/src/main/java/com/adaptris/aws/kinesis/ConnectionFromProperties.java
@@ -14,7 +14,7 @@
     limitations under the License.
 */
 
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import java.io.InputStream;
 import java.util.Properties;
