+++ interlok-aws-kinesis/src/main/java/com/adaptris/aws/kinesis/KinesisStreamProducer.java
@@ -1,4 +1,4 @@
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import java.nio.ByteBuffer;
 import javax.validation.constraints.NotBlank;
