+++ interlok-aws-kinesis/src/test/java/com/adaptris/aws/kinesis/KinesisStreamProducerTest.java
@@ -1,4 +1,4 @@
-package com.adaptris.aws2.kinesis;
+package com.adaptris.aws.kinesis;
 
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
