+++ interlok-aws-sns/build.gradle
@@ -2,15 +2,15 @@ ext {
   componentName = 'Interlok AWS/SNS'
   componentDesc = "Publishing to AWS SNS topics"
   delombokTargetDir = new File("${project.projectDir}/src/main/generated")
-  awsSDKVersion = '2.29.9'
+  awsSDKVersion = '1.12.777'
   jacksonVersion = '2.18.1'
 }
 
 dependencies {
   api (platform("com.fasterxml.jackson:jackson-bom:$jacksonVersion"))
-  api (platform("software.amazon.awssdk:sdk-core:$awsSDKVersion"))
+  api (platform("com.amazonaws:aws-java-sdk-bom:$awsSDKVersion"))
   api ("com.amazonaws:aws-java-sdk-sns")
-  api ("software.amazon.awssdk:sdk-core")
+  api ("com.amazonaws:aws-java-sdk-core")
   // Dependency to platform/BOM
   api ("com.fasterxml.jackson.core:jackson-databind")
   api ("com.fasterxml.jackson.core:jackson-core")
