+++ interlok-aws-sqs/build.gradle
@@ -2,13 +2,13 @@ ext {
   componentName = 'Interlok AWS/SQS'
   componentDesc = "Publishing & Consuming from AWS SQS natively or via their JMS provider"
   delombokTargetDir = new File("${project.projectDir}/src/main/generated")
-  awsSDKVersion = '2.29.9'
+  awsSDKVersion = '1.12.777'
   jacksonVersion = '2.18.1'
 }
 
 dependencies {
-  api ("software.amazon.awssdk:sqs:$awsSDKVersion")
-  api ("software.amazon.awssdk:sdk-core:$awsSDKVersion") {
+  api ("com.amazonaws:aws-java-sdk-sqs:$awsSDKVersion")
+  api ("com.amazonaws:aws-java-sdk-core:$awsSDKVersion") {
     exclude group: "com.fasterxml.jackson.core", module: "jackson-databind"
     exclude group: "com.fasterxml.jackson.core", module: "jackson-core"
     exclude group: "com.fasterxml.jackson.core", module: "jackson-annotations"
